requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
dnspython==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
email-validator==2.1.1
python-dotenv==1.0.1
chromadb
langchain_text_splitters
langchain_huggingface
sentence-transformers==2.7.0
langchain-google-genai
cryptography
bs4
imagekitio
sqlalchemy